feat: add find, count, delete fix: build_tree
Find works - kind of.  It currently chokes when you search for
a value that isn't in the tree.  Still working on that

Build_tree was doing this weird thing where it would add a `nil` node
at the end of the tree when it hit the last element of the array.

count - works

delete - works